+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOAtpPolicyForO365/MSFT_EXOAtpPolicyForO365.psm1
@@ -13,18 +13,9 @@ function Get-TargetResource
         [System.String]
         $Identity = 'Default',
 
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $AllowClickThrough = $true,
-
         [Parameter()]
         [Boolean]
         $AllowSafeDocsOpen = $false,
-
-        [Parameter()]
-        [System.String[]]
-        $BlockUrls = @(),
 
         [Parameter()]
         [Boolean]
@@ -33,16 +24,6 @@ function Get-TargetResource
         [Parameter()]
         [Boolean]
         $EnableSafeDocs = $false,
-
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $EnableSafeLinksForO365Clients = $true,
-
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $TrackClicks = $true,
 
         [Parameter()]
         [ValidateSet('Present')]
@@ -117,16 +98,9 @@ function Get-TargetResource
             $result = @{
                 IsSingleInstance              = "Yes"
                 Identity                      = $AtpPolicyForO365.Identity
-                #DEPRECATED
-                #AllowClickThrough             = $AtpPolicyForO365.AllowClickThrough
                 AllowSafeDocsOpen             = $AtpPolicyForO365.AllowSafeDocsOpen
-                BlockUrls                     = $AtpPolicyForO365.BlockUrls
                 EnableATPForSPOTeamsODB       = $AtpPolicyForO365.EnableATPForSPOTeamsODB
                 EnableSafeDocs                = $AtpPolicyForO365.EnableSafeDocs
-                #DEPRECATED
-                #EnableSafeLinksForO365Clients = $AtpPolicyForO365.EnableSafeLinksForO365Clients
-                #DEPRECATED
-                #TrackClicks                   = $AtpPolicyForO365.TrackClicks
                 Ensure                        = 'Present'
                 ApplicationId                 = $ApplicationId
                 CertificateThumbprint         = $CertificateThumbprint
@@ -180,18 +154,9 @@ function Set-TargetResource
         [System.String]
         $Identity = 'Default',
 
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $AllowClickThrough = $true,
-
         [Parameter()]
         [Boolean]
         $AllowSafeDocsOpen = $false,
-
-        [Parameter()]
-        [System.String[]]
-        $BlockUrls = @(),
 
         [Parameter()]
         [Boolean]
@@ -200,16 +165,6 @@ function Set-TargetResource
         [Parameter()]
         [Boolean]
         $EnableSafeDocs = $false,
-
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $EnableSafeLinksForO365Clients = $true,
-
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $TrackClicks = $true,
 
         [Parameter()]
         [ValidateSet('Present')]
@@ -274,18 +229,6 @@ function Set-TargetResource
     $AtpPolicyParams.Remove('CertificatePassword') | Out-Null
     Write-Verbose -Message "Setting AtpPolicyForO365 $Identity with values: $(Convert-M365DscHashtableToString -Hashtable $AtpPolicyParams)"
 
-    #Deprecated
-    Write-Verbose -Message "Property AllowClickThrough is deprecated."
-    $AtpPolicyParams.Remove('AllowClickThrough') | Out-Null
-
-    #Deprecated
-    Write-Verbose -Message "Property EnableSafeLinksForO365Clients is deprecated."
-    $AtpPolicyParams.Remove('EnableSafeLinksForO365Clients') | Out-Null
-
-    #Deprecated
-    Write-Verbose -Message "Property EnableSafeLinksForO365Clients is deprecated."
-    $AtpPolicyParams.Remove('TrackClicks') | Out-Null
-
     Set-AtpPolicyForO365 @AtpPolicyParams
 }
 
@@ -304,18 +247,9 @@ function Test-TargetResource
         [System.String]
         $Identity = 'Default',
 
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $AllowClickThrough = $true,
-
         [Parameter()]
         [Boolean]
         $AllowSafeDocsOpen = $false,
-
-        [Parameter()]
-        [System.String[]]
-        $BlockUrls = @(),
 
         [Parameter()]
         [Boolean]
@@ -324,16 +258,6 @@ function Test-TargetResource
         [Parameter()]
         [Boolean]
         $EnableSafeDocs = $false,
-
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $EnableSafeLinksForO365Clients = $true,
-
-        #DEPRECATED
-        [Parameter()]
-        [Boolean]
-        $TrackClicks = $true,
 
         [Parameter()]
         [ValidateSet('Present')]
@@ -392,11 +316,6 @@ function Test-TargetResource
     $ValuesToCheck.Remove('CertificateThumbprint') | Out-Null
     $ValuesToCheck.Remove('CertificatePath') | Out-Null
     $ValuesToCheck.Remove('CertificatePassword') | Out-Null
-
-    #DEPRECATED
-    $ValuesToCheck.Remove('AllowClickThrough') | Out-Null
-    $ValuesToCheck.Remove('EnableSafeLinksForO365Clients') | Out-Null
-    $ValuesToCheck.Remove('TrackClicks') | Out-Null
 
     $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `